# BACKEND/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/shipment/{warehouseId}/manifest": {
      "get": {
        "operationId": "ShipmentsController_getManifest",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get manifest/pending shipments for a warehouse",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/shipment/all": {
      "get": {
        "operationId": "ShipmentsController_getAllShipments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all shipments in the system",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/shipment/{warehouseId}/dispatch": {
      "post": {
        "operationId": "ShipmentsController_dispatchShipment",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchShipmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Dispatch a shipment from a warehouse to a hub and agency",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/shipment/summary/{warehouseId}": {
      "get": {
        "operationId": "ShipmentsController_getSummary",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get shipment summary for a warehouse",
        "tags": [
          "Shipments"
        ]
      }
    },
...
    "/agency/{agencyId}/rto": {
      "post": {
        "operationId": "RtoController_createRto",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "trackingId": {
                    "type": "string",
                    "example": "TRK-HYD-20240001"
                  },
                  "reason": {
                    "type": "string",
                    "example": "Customer Refused"
                  },
                  "notes": {
                    "type": "string",
                    "example": "Customer refused to accept the package"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create a Return to Origin (RTO) request",
        "tags": [
          "RTO"
        ]
      },
      "get": {
        "operationId": "RtoController_getAgencyRtos",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all RTOs for an agency",
        "tags": [
          "RTO"
        ]
      }
    },
    "/rto": {
      "get": {
        "operationId": "RtoController_getAllRtos",
        "parameters": [
          {
            "name": "warehouseId",
            "required": false,
            "in": "query",
            "description": "Filter by warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all RTOs (optionally filtered by warehouse)",
        "tags": [
          "RTO"
        ]
      }
    },
    "/rto/{rtoId}": {
      "get": {
        "operationId": "RtoController_getRtoById",
        "parameters": [
          {
            "name": "rtoId",
            "required": true,
            "in": "path",
            "description": "RTO ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get a specific RTO by ID",
        "tags": [
          "RTO"
        ]
      }
    },
    "/rto/{rtoId}/warehouse-inscan": {
      "post": {
        "operationId": "RtoController_warehouseInscan",
        "parameters": [
          {
            "name": "rtoId",
            "required": true,
            "in": "path",
            "description": "RTO ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "In-scan an RTO at the warehouse",
        "tags": [
          "RTO"
        ]
      }
    },
...
    "/hub/{hubId}/inscan": {
      "post": {
        "operationId": "ScanController_inscanAtHub",
        "parameters": [
          {
            "name": "hubId",
            "required": true,
            "in": "path",
            "description": "Transit Hub ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "In-scan a shipment at a transit hub",
        "tags": [
          "Scan"
        ]
      }
    },
    "/hub/{hubId}/outscan": {
      "post": {
        "operationId": "ScanController_outscanFromHub",
        "parameters": [
          {
            "name": "hubId",
            "required": true,
            "in": "path",
            "description": "Transit Hub ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Out-scan a shipment from a transit hub",
        "tags": [
          "Scan"
        ]
      }
    },
    "/hub/{hubId}/scanHistory": {
      "get": {
        "operationId": "ScanController_getScanHistory",
        "parameters": [
          {
            "name": "hubId",
            "required": true,
            "in": "path",
            "description": "Transit Hub ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get scan history for a transit hub",
        "tags": [
          "Scan"
        ]
      }
    },
    "/agency/{agencyId}/inscan": {
      "post": {
        "operationId": "ScanController_agencyInscan",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "In-scan a shipment at a local agency",
        "tags": [
          "Scan"
        ]
      }
    },
    "/agency/{agencyId}/outscan": {
      "post": {
        "operationId": "ScanController_agencyOutscan",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Out-scan a shipment from a local agency",
        "tags": [
          "Scan"
        ]
      }
    },
    "/agency/{agencyId}/scanHistory": {
      "get": {
        "operationId": "ScanController_getAgencyScanHistory",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get scan history for a local agency",
        "tags": [
          "Scan"
        ]
      }
    },
    "/hub/{hubId}/inbound": {
      "get": {
        "operationId": "ReportsController_getInboundForHub",
        "parameters": [
          {
            "name": "hubId",
            "required": true,
            "in": "path",
            "description": "Transit Hub ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get inbound shipments report for a transit hub",
        "tags": [
          "Reports"
        ]
      }
    },
    "/hub/{hubId}/outbound": {
      "get": {
        "operationId": "ReportsController_getOutboundForHub",
        "parameters": [
          {
            "name": "hubId",
            "required": true,
            "in": "path",
            "description": "Transit Hub ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get outbound shipments report for a transit hub",
        "tags": [
          "Reports"
        ]
      }
    },
    "/agency/{agencyId}/inbound": {
      "get": {
        "operationId": "ReportsController_getAgencyInbound",
        "parameters": [
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "description": "Agency ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get inbound shipments for a local agency",
        "tags": [
          "Reports"
        ]
      }
    },
    "/{warehouseId}/outbound": {
      "get": {
        "operationId": "ReportsController_getOutboundShipments",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get outbound shipments from a warehouse",
        "tags": [
          "Reports"
        ]
      }
    },
    "/{warehouseId}/preAlert": {
      "get": {
        "operationId": "ReportsController_getPreAlertReport",
        "parameters": [
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "description": "Warehouse ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get pre-alert shipment report for a warehouse",
        "tags": [
          "Reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "WARE2DOOR API",
    "description": "## WARE2DOOR Logistics API\n\nEnd-to-end logistics platform covering warehouse management, transit hub scanning, last-mile delivery, RTO lifecycle, and super-user administration.\n\n### Authentication\nProtected endpoints require an **`x-role`** header specifying the caller's role.\n\nClick the **Authorize 🔒** button, then enter one of the allowed values:\n\n| Role Value | Access |\n|---|---|\n| `WAREHOUSE` | Warehouse management endpoints |\n| `TRANSIT_HUB` | Transit hub scan & reports endpoints |\n| `LOCAL_AGENCY` | Agency delivery & agent endpoints |\n| `SUPER_USER` | Full administrative access |\n\n> **Public endpoints** (login, verify-email, track, raise-ticket) do **not** require this header.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role-based access header. Allowed values: WAREHOUSE | TRANSIT_HUB | LOCAL_AGENCY | SUPER_USER"
      }
    },
    "schemas": {
      "DispatchShipmentDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "example": "ORD-12345",
            "description": "Order ID to dispatch"
          },
          "hubId": {
            "type": "string",
            "example": "HUB-BLR-001",
            "description": "Transit hub ID to route through"
          },
          "agencyId": {
            "type": "string",
            "example": "AGY-BLR-001",
            "description": "Local agency ID for last-mile delivery"
          }
        },
        "required": [
          "orderId",
          "hubId",
          "agencyId"
        ]
      },
...
      "ScanDto": {
        "type": "object",
        "properties": {
          "trackingId": {
            "type": "string",
            "example": "TRK-HYD-20240001",
            "description": "Tracking ID of the shipment to scan"
          }
        },
        "required": [
          "trackingId"
        ]
      },
...
    }
  },
  "security": [
    {
      "x-role": []
    }
  ]
}